Move FieldsForm Hoc to lib folder
